% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscaleautonet.R
\name{ResautoDownscale}
\alias{ResautoDownscale}
\title{ResautoDownscale}
\usage{
ResautoDownscale(r2, fpredict0, c.grid, ss = 0.2, nepoch = 30,
  cores = 1, thresh = 0.01, ntime = 5)
}
\arguments{
\item{r2}{: Stack for the covariates for downscaling}

\item{fpredict0:}{Starting predictions obtained by GWR or other optimal learners such as XGBoost, residual network}

\item{c.grid:}{Coarsely resolved grid}

\item{ss:}{sampling propotion for independent test}

\item{nepoch:numder}{of epoch for residual network training (default: 30)}

\item{corse:}{Number of CPU cores used for computing (default: 1)}

\item{cores:}{double, dropout rate for the target hidden layer (default 0)}

\item{thresh:}{stopping creation shreshold (default: 0.01)}

\item{ntime:}{maximum number of iterations (default: 5)}
}
\value{
List(performance metrics such as R2, RMSE, and downscaled images)
}
\references{
Lianfa Li et al, 2018, Autoencoder Based Residual Deep Networks for Robust Regression Prediction and Spatiotemporal Estimation, preprint, arXiv:1812.11262 [cs.LG]
}

\description{
This function is the iterative implementation of downscaling with autoencoder based residual network.
}

\author{Lianfa Li \email{lspatial@gmail.com} }

\examples{

\donttest{
    
  tdate="2015-01-01"
  yd=lubridate::yday(tdate)
  print(paste(imgid,":",tdate,sep=""))

tfile=paste("/mnt/server1/GAna/meteomining/chinaimg/imgset/chinaGrdPnt_wp_",tdate,".csv",sep="") 
chinacovs=read.csv(tfile) 
tcols=c("lat","lon","lat2","lon2","lonlat","elevation","julian_day",
        "image_value","pblh")
chinacovsSP=SpatialPointsDataFrame(chinacovs[,c("lon","lat")],chinacovs,
         proj4string=CRS("+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0"))
chinacovsSP_p=spTransform(chinacovsSP,CRSobj=
         CRS("+proj=lcc +lat_1=30 +lat_2=62 +lat_0=0 +lon_0=105 +x_0=0 +y_0=0 +ellps=krass +units=m +no_defs"))
gridded(chinacovsSP_p)=TRUE
chinastk=stack(chinacovsSP_p)

gldas_re2_cn_p=raster(tpath)

ares=ResautoDownscaleOnline(r2,fpredict0,c.grid,ss= 0.2, cores= 5, thresh = 0.01,ntime=6)

 }
}






