\name{resautonet-package}
\alias{resautonet-package}
\alias{resautonet}
\docType{package}
\title{
\packageTitle{resautonet}
}
\description{
\packageDescription{resautonet}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{resautonet}
\packageIndices{resautonet}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{resautonet}

Maintainer: \packageMaintainer{resautonet}
}
\references{
~~ Literature or other references for background information ~~
}
~~ ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
